refactor Metadata into Posts collection, and correct Meteor->React reactiveness!
DID YOU KNOW?

ALthough "reactivity" is central to both Metero and Reactjs, they handle the concept a bit differently. In Meteor, some data structures are 'reactive data sources,' these include Mongo cursors and session objects. When a reactive data source is nested in a Tracker.autorun() callback function, the callback function will rerun, or autorun, when the data source changes. In Reacjs, the views react to the state and changes in state. However, the state object is not reactive in itself. In other words, when using React, you have to handle the changes in data model and don't worry about the view, while in Meteor, it automates both. Toward this goal, Reactjs exposures a range of lifecycle funcitons to update the state and props as needed.

When integrating Meteor with React, you have to be careful what sort of data source is handed to React and how React handles it. Case in point, we were running into a bug in which React was not rendering a document fetch correctly. The problem was that the data (from the document fetch) is set in setInitialState, and never updated again. Since mongo operations are asynchronous, the data will update with results as they return. The fix was to update the state in componentWillReceiveProps(). Interesting to note, when the data handed to React was a cursor, the state update was not necessary because the cursor points to the current set of data.